Check carry propagation (NOT FOR PRODUCTION)
This patch is meant to check that the invariants of carry propagation
are respected.  The invariants are such that operands are allowed to be
added twice before carry propagation must happen (typically during a
multiplication).

For instance:

    x = constant1; // okay       (1x)
    y = constant2; // okay       (1x)
    a = x + y;     // still okay (2x)
    b = a + x;     // still okay (3x)
    c = a + a;     // OVERFLOW   (4x)

The patch works by instrumenting fundamental field operations.  We add a
"carry limb" to field elements, which trace how far we've taken the
addition chains.  If we exceed 3, that's an overflow.

- fe_0(), fe_1() and fe_frombytes() set the carry limb to 1;

- fe_copy() copy the carry limb.

- fe_cswap() and fe_ccopy() take the maximum of the carry limb, and copy
  it to the relevant operands.  This guarantees an upper bound.

- fe_mul_small(), fe_mul(), and fe_sq() actually perform carry
  propagation, so they reset the carry limb to 1.

- fe_add() and fe_neg() add the carry limbs, and asserts that the sum
  does not exceed 3.  (They also assert that no value is negative, in
  case we failed to initialise some carry limb.)  Since addition and
  subtraction are the only operation that march us towards overflow,
  they're the only ones that need asserts.

  Note that zeroes are interpreted as if they were ones.  This allow us
  to leave the constants alone (C initialise unspecified elements with 0
  when the array is not on the stack—and constants are all declared
  static).

If the tests do not fail an assert, that's pretty strong evidence that
the code does not have any limb overflow problem.  This evidence is even
stronger when we pass the full test suite, with the sanitisers.